Se crearo los métodos obtenerOrdenPorId(String idOrden) y cancelarOrden(String idOrden)

+++ src/main/java/co/edu/uniquindio/uniEventos/dto/PagoDTO.java
@@ -1,11 +1,13 @@
 package co.edu.uniquindio.uniEventos.dto;
 
+import jakarta.validation.constraints.NotBlank;
+
 public record PagoDTO(
         String moneda,
         String codigoAutorizacion,
         String tipoPago,
         String detalleEstado,
-        double valorTransaccion,
+        float valorTransaccion,
         String estado
 
 ) {
